third_party/WebKit/LayoutTests/http/tests/misc/onload-detach-during-csp-frame-src-none-expected.txt

CONSOLE ERROR: Refused to frame 'data:,hello' because it violates the following Content Security Policy directive: "frame-src 'none'".

PASS
